Achieve Larastan level 10 with zero errors (was 386)
Type-level fixes across ~50 files (no @phpstan-ignore, no baseline):
- @property docblocks + relation generics (HasMany/BelongsTo<..., $this>) on models
- typed returns/params, PHPDoc generics on collections/arrays
- narrowed mixed (config/data_get/Get-Set/HTTP json) via is_* guards & casts
- added missing match default arms (behavior-preserving)

Real bugs fixed along the way:
- WriterExporter 'Durum' column always showed 'Tamamlandı' (tested $record object,
  always truthy) -> now reads the is_finished state
- env() calls moved to config() (safe under config:cache); added
  services.open_exchange_rates / services.metals keys so values still resolve
- TasksBoardBoardPage: Task::find(mixed) could return a Collection -> guarded to ?Task

Verified: phpstan level 10 clean, pint passes, app boots, 11 key admin pages render.

+++ app/Filament/Clusters/Library/Resources/BookResource/Widgets/BooksOverview.php
@@ -38,6 +38,6 @@ class BooksOverview extends BaseWidget
 
     private function getPageCount(): int
     {
-        return Book::sum('page_count');
+        return (int) Book::sum('page_count');
     }
 }
